Removed Guard
Necessary since this gem is no longer maintained and is often the source of multiple Ruby warnings when running the test suite locally. As an alternative, I've started using link:https://github.com/AlexB52/retest[Retest] which is unobtrusive and doesn't require a configuration (although this might change in the future).

Milestone: patch