Added registration directives
Necessary to allow dependencies to be registered via the following directives:

* `cache`: Ensures the dependency is cached so only a single instance of the dependency is created when resolved. This has been and is the default behavior.
* `fresh`: Ensures a dependency is never cached so a fresh instance is created each time it's resolved.

This means the following is equivalent to existing behavior:

[source,ruby]
----
register(:demo) { Object.new }
register(:demo, as: :cache) { Object.new }
----

Use of `as: :cache` is identical to not supplying a value for `as` so doesn't need to be used since it's redundant. What _is_ new is you can force your dependency to be a new instance when resolved by using `:fresh`. Example:

[source,ruby]
----
register(:demo, as: :fresh) { Object.new }
----

The above will ensure a new (fresh) instance of `Object` is created each time it is revolved instead of always returning the same instance.

The implementation was updated to use a tuple for all registered values in the form of a value and directive (i.e. cache or fresh) instead having a single value. This is a structural change which is why all specs were updated in order to ensure the entire test suite remains green. Care was taken, in the `Builder`, to ensure you can still iterate over a container's key/value pairs minus directives (i.e. `:cache`, `:fresh`) since that is internal information.

Milestone: minor